+++ deluge/ui/web/js/deluge-all/add/UrlWindow.js
@@ -72,6 +72,7 @@ Deluge.add.UrlWindow = Ext.extend(Deluge.add.Window, {
         } else {
             deluge.client.web.download_torrent_from_url(url, cookies, {
                 success: this.onDownload,
+                failure: this.onDownloadFailed,
                 scope: this,
                 torrentId: torrentId,
             });
@@ -85,10 +86,23 @@ Deluge.add.UrlWindow = Ext.extend(Deluge.add.Window, {
     onDownload: function(filename, obj, resp, req) {
         deluge.client.web.get_torrent_info(filename, {
             success: this.onGotInfo,
+            failure: this.onDownloadFailed,
             scope: this,
             filename: filename,
             torrentId: req.options.torrentId,
         });
+    },
+
+    onDownloadFailed: function(obj, resp, req) {
+        Ext.MessageBox.show({
+            title: _('Error'),
+            msg: _('Failed to download torrent'),
+            buttons: Ext.MessageBox.OK,
+            modal: false,
+            icon: Ext.MessageBox.ERROR,
+            iconCls: 'x-deluge-icon-error',
+        });
+        this.fireEvent('addfailed', req.options.torrentId);
     },
 
     onGotInfo: function(info, obj, response, request) {